feat: /hooks add custom hook for modalPathname
