+++ src/components/Header/Header.jsx
@@ -11,24 +11,17 @@ import {
 	logOut,
 } from '../../redux/reducers/authentification';
 import logo from '../../assets/upTalent.png'
+import { useModalPathname } from '../../hooks/useModalPathname';
 
 export const Header = () => {
 	const dispatch = useDispatch();
 	const isTalent = useSelector(getIsAuth);
 	const authTalent = useSelector(getFirstName);
 	const authTalentId = useSelector(getAuthTalentId);
+	const modalPathname = useModalPathname();
 
 	const [dropdownMenu, setDropdownMenu] = useState(null);
-
-	const location = useLocation();
 	const navigate = useNavigate();
-
-	const modalPathname = path => {
-		navigate({
-			pathname: `${location.pathname}/${path}`,
-			search: location.search,
-		});
-	};
 
 	const handleClick = event => {
 		setDropdownMenu(event.currentTarget);
