+++ src/components/Home/Home.jsx
@@ -11,6 +11,7 @@ import {
 } from '../../redux/reducers/authentification';
 import { TalentsSlider } from './components/TalentsSlider';
 import { ProofSlider } from './components/ProofSlider/ProofSlider';
+import { useModalPathname } from '../../hooks/useModalPathname';
 
 export const Home = () => {
 	const isTalent = useSelector(getIsAuth);
@@ -18,13 +19,8 @@ export const Home = () => {
 	const systemMessage = useSelector(state => state.systemMessage.messageText);
 
 	const navigate = useNavigate();
-	const location = useLocation();
+	const modalPathname = useModalPathname();
 
-	const modalPathname = path => {
-		navigate(`${location.pathname}/${path}`, {
-			state: { from: location.pathname },
-		});
-	};
 	return (
 		<>
 			<div className={styles.container}>
